improve kafka pipeline

+++ money-and-account-management-service/src/main/java/com/example/bank/dto/CreditResponseMessage.java
@@ -13,5 +13,5 @@ public class CreditResponseMessage implements Serializable{
 
     private String accountNumber;
 
-    private Float money;
+    private Long money;
 }
